fix: Improve defaults type at schema declaration
When building a schema we're currently assigning `TDefaults` to
`Partial<TProperties>` which is breaking the type safety for creating
new documents provided by `DocumentForInsertWithoutDefaults`. This is
because that type util (correctly) removes supplied default values for
the model from the required properties when inserting a document, but
because _all_ properties from `TProperties` are included in the type
definition, it results in no keys remaining.

This wasn't caught in testing because rather than supplying our
`Model<>` type with the default value returned when creating a schema,
we're directly gathering the default type via `typeof DEFAULTS`. This
commit also updates the tests to pull the type from the returned schema
type signature.