feat: Add `Model.exists(filter)` query helper
There's a little bit of nuance to getting the most performance out of a
query when you only care checking if there exists a row matching a given
query. This helper tries its best to query the DB in the most performant
way, and returns a boolean result.

Performance Notes:
- The query used is a `findOne`. The query engine can then start scanning and stop
  as soon as it matches a single document.

- Project out a single column from the filter criteria (or project `_id`
  if there is no filter). In this way we return a minimum of data, while
  potentially allowing this to be a "covered" query if there is an index
  that includes all the keys in the filter.