feat: `null` type
While storing `null` literals in Mongo has some surprising semantics
(and this author thinks the practice should be strongly discouraged), if
you are trying to add a schema to a collection that already has literal
`null` values stored we should make it possible to type that correctly.

I've added the raw `types.null()`, but have left it a bit intentionally
difficult to use: naively using it results in a column that may _only_
contain the `null` value, so it should typically be combined with other
types. I considered adding a helper like `types.nullable(types.number())`
but I felt that adding that syntactic sugar would make folks more likely
to use it.